+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # 1.0.12 - 5 Apr 2024
+Improvement:
+- export `InferContext` and `InferHandler`
+
+Bug fix:
+- remove accidental `console.log` in `compile`
+
+# 1.0.12 - 5 Apr 2024
 Feature:
 - add `InferContext`
 
 Bug fix:
 - returning null with response validation cause error
-- 
 
 # 1.0.11 - 2 Apr 2024
 Bug fix: